+++ src/ByteDev.Cmd.UnitTests/Tables/TablePositionTests.cs
@@ -1,7 +1,8 @@
 ﻿using System;
+using ByteDev.Cmd.Tables;
 using NUnit.Framework;
 
-namespace ByteDev.Cmd.UnitTests
+namespace ByteDev.Cmd.UnitTests.Tables
 {
     [TestFixture]
     public class TablePositionTests
@@ -12,19 +13,19 @@ namespace ByteDev.Cmd.UnitTests
             [Test]
             public void WhenColumnNumberLessThanZero_ThenThrowException()
             {
-                Assert.Throws<ArgumentOutOfRangeException>(() => new TablePosition(-1, 0));
+                Assert.Throws<ArgumentOutOfRangeException>(() => new CellPosition(-1, 0));
             }
 
             [Test]
             public void WhenRowNumberLessThanZero_ThenThrowException()
             {
-                Assert.Throws<ArgumentOutOfRangeException>(() => new TablePosition(0, -1));
+                Assert.Throws<ArgumentOutOfRangeException>(() => new CellPosition(0, -1));
             }
 
             [Test]
             public void WhenColumnAndRowValid_ThenSetProperties()
             {
-                var sut = new TablePosition(0, 1);
+                var sut = new CellPosition(0, 1);
 
                 Assert.That(sut.Column, Is.EqualTo(0));
                 Assert.That(sut.Row, Is.EqualTo(1));
@@ -37,7 +38,7 @@ namespace ByteDev.Cmd.UnitTests
             [Test]
             public void WhenCalled_ThenReturnsStringRepresentation()
             {
-                var sut = new TablePosition(0, 1);
+                var sut = new CellPosition(0, 1);
 
                 var result = sut.ToString();