+++ src/ByteDev.Cmd/Tables/Borders/BorderSimple.cs
@@ -1,0 +1,38 @@
+﻿namespace ByteDev.Cmd.Tables.Borders
+{
+    /// <summary>
+    /// Represents a simple border style.
+    /// </summary>
+    public class BorderSimple : IBorderStyle
+    {
+        /// <summary>
+        /// Character for horizontal line.
+        /// </summary>
+        public char HorizontalLine => '-';
+
+        /// <summary>
+        /// Character for vertical line.
+        /// </summary>
+        public char VerticalLine => '|';
+
+        /// <summary>
+        /// Character for left top corner.
+        /// </summary>
+        public char LeftTop => '+';
+
+        /// <summary>
+        /// Character for right top corner.
+        /// </summary>
+        public char RightTop => '+';
+
+        /// <summary>
+        /// Character for left bottom corner.
+        /// </summary>
+        public char LeftBottom => '+';
+
+        /// <summary>
+        /// Character for right bottom corner.
+        /// </summary>
+        public char RightBottom => '+';
+    }
+}