+++ src/ByteDev.Cmd/Tables/CellPosition.cs
@@ -1,11 +1,11 @@
 ﻿using System;
 
-namespace ByteDev.Cmd
+namespace ByteDev.Cmd.Tables
 {
     /// <summary>
-    /// Represents a cell position in <see cref="T:ByteDev.Cmd.Table" />.
+    /// Represents a cell position in <see cref="T:ByteDev.Cmd.Tables.Table" />.
     /// </summary>
-    public struct TablePosition
+    public struct CellPosition
     {
         /// <summary>
         /// Column number position.
@@ -18,13 +18,13 @@ namespace ByteDev.Cmd
         public int Row { get; }
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="T:ByteDev.Cmd.TablePosition" /> class.
+        /// Initializes a new instance of the <see cref="T:ByteDev.Cmd.Tables.CellPosition" /> class.
         /// </summary>
         /// <param name="column">Column number position. First position is zero.</param>
         /// <param name="row">Row number position. First position is zero.</param>
         /// <exception cref="T:System.ArgumentOutOfRangeException"><paramref name="column" /> cannot be less than zero.</exception>
         /// <exception cref="T:System.ArgumentOutOfRangeException"><paramref name="row" /> cannot be less than zero.</exception>
-        public TablePosition(int column, int row)
+        public CellPosition(int column, int row)
         {
             if (column < 0)
                 throw new ArgumentOutOfRangeException(nameof(column), column, "Column number cannot be less than zero.");
@@ -37,9 +37,9 @@ namespace ByteDev.Cmd
         }
 
         /// <summary>
-        /// Returns a string representation of <see cref="T:ByteDev.Cmd.TablePosition" />.
+        /// Returns a string representation of <see cref="T:ByteDev.Cmd.Tables.CellPosition" />.
         /// </summary>
-        /// <returns>String representation of <see cref="T:ByteDev.Cmd.TablePosition" />.</returns>
+        /// <returns>String representation of <see cref="T:ByteDev.Cmd.Tables.CellPosition" />.</returns>
         public override string ToString()
         {
             return $"{Column}x{Row}";